select: prevent a race condition in locking Select.choices
When multiple threads were selecting on the same channels,
the following race condition could occur in Select::maybe_try_select:

Provided
- Two treads: `A` and `B` and two channels `chan1` and `chan2`
- self.choices for thread `A` is [`chan1`, `chan2`]
- self.choices for thread `B` is [`chan2`, `chan1`]

In a for loop in Select::maybe_try_select, the following happens:
`A` tries to lock `chan1`, this succeeds
`B` tries to lock `chan2`, this succeeds
`A` tries to lock `chan2`, this blocks, because `B` has locked it
`B` tries to lock `chan1`, this blocks, because `A` has locked it
*deadlock!*

This can be prevented by making sure that the different threads all
lock the channels in the same order. The order of HashMap::iter()
is arbitrary, so the bug should be fixed when we instead use a BTreeMap.

Fixes #5.